--- Include/Reflection/Static/BaseType.h
@@ -1,25 +1,0 @@
-// Copyright 2015-2021 Piperift - All rights reserved
-#pragma once
-
-#include "PCH.h"
-
-#include "Reflection/TypeId.h"
-
-
-namespace Rift::Refl
-{
-	/** Smallest reflection type */
-	class CORE_API BaseType
-	{
-	protected:
-		TypeId id;
-
-
-	protected:
-		BaseType() = default;
-
-	public:
-		BaseType(const BaseType&) = delete;
-		BaseType& operator=(const BaseType&) = delete;
-	};
-}    // namespace Rift::Refl
