+++ Include/Reflection/Static/TProperty.h
@@ -4,8 +4,8 @@
 #include "Events/Function.h"
 #include "Reflection/ReflectionTags.h"
 #include "Reflection/Runtime/TPropertyHandle.h"
+#include "Reflection/Static/DataType.h"
 #include "Reflection/Static/Property.h"
-#include "Reflection/Static/Type.h"
 #include "Strings/Name.h"
 
 #include <functional>
