+++ Include/Reflection/Static/Type.h
@@ -3,86 +3,23 @@
 
 #include "PCH.h"
 
-#include "Containers/Array.h"
-#include "Containers/Map.h"
-#include "CoreTypes.h"
-#include "Reflection/ReflectionTags.h"
-#include "Reflection/Static/BaseType.h"
-#include "Reflection/Static/Property.h"
-#include "Strings/String.h"
+#include "Reflection/TypeId.h"
 
 
 namespace Rift::Refl
 {
-	using PropertyMap = TMap<Name, Property*>;
-
-	/** Smallest reflection type that contains all basic class or struct data */
-	class CORE_API DataType : public BaseType
+	/** Smallest reflection type */
+	class CORE_API Type
 	{
-		template <typename T, typename Parent, typename TType, ReflectionTags tags>
-		friend struct TTypeBuilder;
-
 	protected:
-		Name name;
-		ReflectionTags tags = ReflectionTags::None;
-
-		DataType* parent = nullptr;
-		TArray<DataType*> children;
-
-		PropertyMap properties{};
+		TypeId id;
 
 
 	protected:
-		DataType() = default;
+		Type() = default;
 
 	public:
-		DataType(const DataType&) = delete;
-		DataType& operator=(const DataType&) = delete;
-		virtual ~DataType()
-		{
-			for (auto& it : properties)
-			{
-				delete it.second;
-			}
-		}
-
-		/** Type */
-		const Name& GetName() const
-		{
-			return name;
-		}
-		const String& GetSName() const
-		{
-			return GetName().ToString();
-		}
-
-		bool HasTag(ReflectionTags tag) const
-		{
-			return (tags & tag) > 0;
-		}
-
-		bool IsChildOf(const DataType* other) const;
-
-		template <typename T>
-		bool IsChildOf() const
-		{
-			return IsChildOf(GetType<T>());
-		}
-
-		bool IsParentOf(const DataType* other) const
-		{
-			return other && other->IsChildOf(this);
-		}
-
-
-		/** Properties */
-		const Property* FindProperty(const Name& propertyName) const;
-		void GetOwnProperties(PropertyMap& outProperties) const;
-		void GetAllProperties(PropertyMap& outProperties) const;
-
-
-	protected:
-		void __GetAllChildren(TArray<DataType*>& outChildren);
-		DataType* __FindChild(const Name& className) const;
+		Type(const Type&) = delete;
+		Type& operator=(const Type&) = delete;
 	};
 }    // namespace Rift::Refl
