+++ Src/Object/BaseStruct.cpp
@@ -2,9 +2,5 @@
 
 #include "Object/BaseStruct.h"
 
-#include "Object/Struct.h"
-#include "Reflection/Static/Struct.h"
-
-
 namespace Rift
 {}    // namespace Rift
