+++ Src/Reflection/Static/Property.cpp
@@ -3,8 +3,8 @@
 #include "Reflection/Static/Property.h"
 
 #include "Reflection/Runtime/PropertyHandle.h"
-#include "Reflection/Static/Class.h"
-#include "Reflection/Static/Struct.h"
+#include "Reflection/Static/ClassType.h"
+#include "Reflection/Static/StructType.h"
 
 
 namespace Rift::Refl
